requirements.txt
streamlit
numpy
matplotlib
scikit-learn
faiss-cpu
sentence-transformers
transformers
torch
pillow
pytesseract
python-dotenv
openai
nltk
rouge
PyMuPDF